Extract class to file
